fix init 😎
