+++ NAMESPACE
@@ -1,4 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
+export(ok_proj6)
+export(proj_trans_generic)
 useDynLib(PROJ, PROJ_proj_trans_generic)
 useDynLib(PROJ, PROJ_set_data_dir)
